ensemble: genericization wave — public photometry + surveys.roman modules, flux_unit metadata
- kl_pipe/photometry.py: AB<->uJy, fnu->flambda, power-law band interpolation,
  unit constants (UJY_TO_CGS/CGS_FNU_TO_UJY consolidated)
- kl_pipe/surveys/roman.py: all HLWAS depth/limit/band constants + the
  depth-referenced SNR and measurement-error helpers, provenance intact
- noise.py: generic gaussian_matched_filter_compactness(r50, cosi, psf_fwhm)
- population.py slims to selection+paint, calls the public tools; adapter
  import cycles gone
- flux_unit metadata on ImageObs/GrismObs, set in catalog mode; naming-rule
  paragraph in units doc
- pinned conversion/depth tests move verbatim to generic-tier
  test_photometry.py + test_surveys_roman.py (21 moved, 11 new)
