+++ kl_pipe/surveys/__init__.py
@@ -1,0 +1,4 @@
+"""
+Published survey parameters (depths, flux limits, band definitions),
+one module per survey.
+"""
